+++ CHANGELOG.rst
@@ -15,6 +15,18 @@ Change Log
 
 Unreleased
 ----------
+[4.6.7]
+-------
+feat: filter courses from API Response of SAPSF to store in the APIResponseRecord table
+
+[4.6.6]
+-------
+chore: orchestrator exception handling and submission refinements
+
+[4.6.5]
+-------
+feat: Added logs for Degreed2 client
+
 [4.6.4]
 -------
 chore: updating sso orchestrator self service api endpoints